js/src/xpconnect/tests/chrome/bug503926.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=503926
-->
<window title="Mozilla Bug 503926"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=503926"
     target="_blank">Mozilla Bug 503926</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[
      var passed = false;
      var obj = { QueryInterface: function() { passed = true; } }
      try { document.documentElement.appendChild(obj); } catch (e) {}
      var isDialog = location.hash != 'iframe';
      var outer = isDialog ? opener.wrappedJSObject : top.wrappedJSObject;
      outer.ok(passed, "chrome/chrome test passed");
      if (isDialog)
        close();
  ]]>
  </script>
</window>
